+++ praw/const.py
@@ -123,6 +123,7 @@ API_PATH = {
     'wiki_edit':              'r/{subreddit}/api/wiki/edit/',
     'wiki_page':              'r/{subreddit}/wiki/{page}',
     'wiki_page_editor':       'r/{subreddit}/api/wiki/alloweditor/{method}',
+    'wiki_page_revisions':    'r/{subreddit}/wiki/revisions/{page}',
     'wiki_page_settings':     'r/{subreddit}/wiki/settings/{page}',
     'wiki_pages':             'r/{subreddit}/wiki/pages/',
     'wiki_revisions':         'r/{subreddit}/wiki/revisions/'}